+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/MultisigRpc.scala
@@ -43,7 +43,7 @@ trait MultisigRpc { self: Client =>
         JsString(account)
       ) ++ addressType.map(Json.toJson(_)).toList
 
-    self.version.flatMap { case V24 | V23 | V22 | Unknown =>
+    self.version.flatMap { case V24 | V23 | Unknown =>
       bitcoindCall[MultiSigResultPostV20](
         "addmultisigaddress",
         params,
@@ -86,7 +86,7 @@ trait MultisigRpc { self: Client =>
       addressType: AddressType,
       walletNameOpt: Option[String] = None
   ): Future[MultiSigResult] = {
-    self.version.flatMap { case V24 | V23 | V22 | Unknown =>
+    self.version.flatMap { case V24 | V23 | Unknown =>
       bitcoindCall[MultiSigResultPostV20](
         "createmultisig",
         List(
