+++ testkit/src/main/scala/org/bitcoins/testkit/node/fixture/NodeConnectedWithBitcoind.scala
@@ -2,7 +2,6 @@ package org.bitcoins.testkit.node.fixture
 
 import org.bitcoins.node.{NeutrinoNode, Node}
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
-import org.bitcoins.rpc.client.v22.BitcoindV22RpcClient
 
 /** Gives us a fixture that has a Neutrino node connected with the bitcoind
   * instance
@@ -26,11 +25,6 @@ case class NeutrinoNodeConnectedWithBitcoinds(
     node: NeutrinoNode,
     bitcoinds: Vector[BitcoindRpcClient]
 ) extends NodeConnectedWithBitcoinds
-
-case class NeutrinoNodeConnectedWithBitcoindV22(
-    node: NeutrinoNode,
-    bitcoind: BitcoindV22RpcClient
-) extends NodeConnectedWithBitcoind
 
 case class NeutrinoNodeNotConnectedWithBitcoinds(
     node: NeutrinoNode,
